refactor: LineEntity 추가
