+++ src/main/java/wooteco/subway/entity/LineEntity.java
@@ -1,17 +1,17 @@
-package wooteco.subway.dto.request;
+package wooteco.subway.entity;
 
-public class LineSaveRequest {
+public class LineEntity {
 
     private final String name;
     private final String color;
 
-    private LineSaveRequest(String name, String color) {
+    private LineEntity(String name, String color) {
         this.name = name;
         this.color = color;
     }
 
-    public static LineSaveRequest of(String name, String color) {
-        return new LineSaveRequest(name, color);
+    public static LineEntity of(String name, String color) {
+        return new LineEntity(name, color);
     }
 
     public String getName() {
